Combined ReminderForm with ReminderCrud
Did the same thing with reminders as was done with notes in the last commit
